requirements.txt
argopt==0.6.1
asgiref==3.2.10
coverage==5.3
Django==3.1.2
django-dotenv==1.4.2
django-polymorphic==3.0.0
dj-database-url==0.5.0
fame==1.2.0
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
tabulate==0.8.7
tqdm==4.51.0
django-heroku==0.3.1
whitenoise==5.2.0
reportlab==3.5.56
